Better deal with scanner errors
The block scanner was quietly erroring on large blocks. It turns
out that when the scanner returns False, we should check scanner.Err()
to see whether we got an error or reached the end.

I increased the buffer size for the block scanner. I haven't seen
any reason to increase it for the order scanner, but still added the
error check to avoid silent errors.